feat: replace ownable with default admin role
Use `_grantRole` vs the suggestion of `_setupRole`
because the natspec of `AccessControl` contract
mentions `_setupRole` as deprecated (link below).

Removing Ownable and switching entirely to AccessControl removes the
need for specific methods for granting and revoking roles.  However
we keep those methods in this commit and remove them in later
commits, in order to ensure that each commit is tightly scoped and
contains only a single logical change.

https://github.com/OpenZeppelin/openzeppelin-contracts/blob/815bfbf38395e0c836d09745ba69ae52844e3d8c/contracts/access/AccessControl.sol#L203