feat: remove custom token issuer role
Instead, use the `AccessControl` contract to unify access control in just this one mechanism, per our auditors' recommendation. This allows us to remove the function and event used for access until now.

Within the above, a failed test was introduced as the `broker` role was expected to have token issuer access. There is no specific place where a broker becomes a token issuer before that, not in the contract or test. The flow previously is: the contract gets deployed with `bridgeAdmin` (now `tokenIssuer`) as a token issuer. At one later point, the token issuer role is given to `admin`. But never to `broker`.

Fix failing test by calling token issuence from a caller that has the token issuer role.