feat: constructor assigns roles to accounts
This allows us to deploy & assign roles in just one call.
It can be useful not to do the above in two calls,
especially when deploying from a multi-sig.

We also stop assigning the deployer as a default admin for the contract.
The admin now needs to be specified in the constructor arguments. This
allows for more flexibility when deploying.